formlayout fix

+++ lib/core/ui/layout/widgets/co_form_layout_widget.dart
@@ -964,7 +964,7 @@ class RenderFormLayoutWidget extends CoLayoutRenderBox
           } else {
             topAnchor.position = ((size.height - maxSize.height) / 2).round();
           }
-          bottomAnchor.position = topAnchor.position + maxSize.height;
+          bottomAnchor.position = (topAnchor.position + maxSize.height) as int;
         } else {
           topAnchor.position = 0;
           bottomAnchor.position = size.height.round();
